+++ src/options.c
@@ -1,4 +1,3 @@
-#include <stdint.h>
 #include <stdlib.h>
 #include <string.h>
 #include "options.h"
@@ -498,6 +497,7 @@ static const OptionDesc option_desc[] = {
     STR_OPT("statusline-left", G(statusline_left), validate_statusline_format, NULL),
     STR_OPT("statusline-right", G(statusline_right), validate_statusline_format, NULL),
     BOOL_OPT("syntax", C(syntax), syntax_changed),
+    FSIZE_OPT("syntax-line-limit", G(syntax_line_limit), NULL),
     BOOL_OPT("tab-bar", G(tab_bar), redraw_screen),
     UINT8_OPT("tab-width", C(tab_width), 1, TAB_WIDTH_MAX, redraw_buffer),
     UINT_OPT("text-width", C(text_width), 1, TEXT_WIDTH_MAX, NULL),